% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA2022-package.R
\docType{data}
\name{causetext2022}
\alias{causetext2022}
\title{Causes for the WHO 2022 instrument}
\format{
A data frame with 67 rows and 3 columns.
}
\description{
This is the table of VA causes and their codes for the WHO 2022 instrument.
}
\examples{

data(causetext2022)

}
\keyword{datasets}
