% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA2022.plot.R
\name{CSMF.interVA2022}
\alias{CSMF.interVA2022}
\title{Summarize population level cause-specific mortality fraction as InterVA2022 suggested.}
\usage{
CSMF.interVA2022(va)
}
\arguments{
\item{va}{The list of va object to summarize.}
}
\value{
\item{dist.cod}{The cause-specific mortality fraction (including undetermined category).}
}
\description{
The function takes input of a list of va object and calculates the
cause-specific mortality fraction. It only calculates CSMF5 as aggregation of up to the third largest causes.
}
\examples{

\dontrun{
data(va2022)

sample.output <- InterVA2022(va2022, HIV = "h", Malaria = "v", Covid = "v",
       write=TRUE, directory = tempdir(), filename = "VA2022_result",
       output = "extended", append = FALSE)
## Get CSMF without plots
csmf <- CSMF.interVA2022(sample.output$VA2022)
}

}
\seealso{
\code{\link{CSMF5}}
}
\author{
CDC Foundation
}
\keyword{interVA}
