% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA2022.plot.R
\name{InterVA2022.plot}
\alias{InterVA2022.plot}
\title{Plot an individual-level distribution of va probabilities.}
\usage{
InterVA2022.plot(
  va,
  type = "bar",
  title = "Top CSMF Distribution",
  min.prob = 0.01,
  ...
)
}
\arguments{
\item{va}{A va object}

\item{type}{An indicator of the type of chart to plot.  "pie" for pie chart;
"bar" for bar chart.}

\item{title}{A character string for the title of the CSMF plot.}

\item{min.prob}{The minimum probability that is to be plotted in bar chart,
or to be labeled in pie chart.}

\item{...}{Arguments to be passed to/from graphic function
\code{\link[graphics]{barplot}}, \code{\link[graphics]{pie}}, and more
graphical paramters (see \code{\link[graphics]{par}}). They will affect the
main title, size and font of labels, and the radius of the pie chart.}
}
\description{
The function takes an input of a single va object and produces a summary plot
for it.
}
\examples{

\dontrun{
data(va2022)
#' sample.output <- InterVA2022(va2022, HIV = "h", Malaria = "v", Covid = "v",
    write = FALSE, directory = tempdir(), filename = "VA2022_result",
    output = "extended", append = FALSE)

## Individual level summary using pie chart
InterVA2022.plot(sample.output$VA2022[[1]], type = "pie", min.prob = 0.01,
    main = "1st sample VA analysis using pie chart", clockwise = FALSE,
    radius = 0.6, cex = 0.6, cex.main = 0.8)

## Individual level summary using bar chart
InterVA2022.plot(sample.output$VA2022[[1]], type = "bar", min.prob = 0.01,
    main = "2nd sample VA analysis using bar chart", cex.main = 0.8)
}

}
\seealso{
\code{\link{CSMF5}}
}
\keyword{InterVA}
