% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA2022.R
\name{InterVA2022}
\alias{InterVA2022}
\title{Provide InterVA2022 analysis on the data input.}
\usage{
InterVA2022(
  Input,
  HIV,
  Malaria,
  Covid,
  write = TRUE,
  directory = NULL,
  filename = "VA2022_result",
  output = "classic",
  append = FALSE,
  groupcode = FALSE,
  sci = NULL,
  ...
)
}
\arguments{
\item{Input}{A matrix input, or data read from csv files in the same format
as required by InterVA2022.}

\item{HIV}{An indicator of the level of prevalence of HIV. The input should
be one of the following: "h"(high),"l"(low), or "v"(very low).}

\item{Malaria}{An indicator of the level of prevalence of Malaria. The input
should be one of the following: "h"(high),"l"(low), or "v"(very low).}

\item{Covid}{An indicator of the level of prevalence of Covid19. The input
should be one of the following: "h"(high),"l"(low), or "v"(very low).}

\item{write}{A logical value indicating whether or not the output (including
errors and warnings) will be saved to file.  If the value is set to TRUE, the
user must also provide a value for the parameter "directory".}

\item{directory}{The directory to store the output from InterVA2022. It should
either be an existing valid directory, or a new folder to be created. If no
path is given and the parameter for "write" is true, then the function stops
and and error message is produced.}

\item{filename}{The file name the user wish to save the output. No extension
needed. The output is in .csv format by default.}

\item{output}{"classic": The same delimited output format as InterVA2022; or
"extended": delimited output followed by full distribution of cause of
death probability.}

\item{append}{A logical value indicating whether or not the new output
should be appended to the existing file.}

\item{groupcode}{A logical value indicating whether or not the group code
will be included in the output causes.}

\item{sci}{A data frame that contains the symptom-cause-information (aka
Probbase) that InterVA uses to assign a cause of death.}

\item{...}{not used}
}
\value{
\item{ID }{ identifier from batch (input) file} \item{MALPREV
}{ selected malaria prevalence} \item{HIVPREV }{ selected HIV prevalence}
\item{COVIDPREV }{ selected Covid19 prevalence}
\item{PREGSTAT }{most likely pregnancy status} \item{PREGLIK }{ likelihood of
PREGSTAT} \item{PRMAT }{ likelihood of maternal death} \item{INDET
}{ indeterminate outcome} \item{CAUSE1 }{ most likely cause} \item{LIK1 }{
likelihood of 1st cause} \item{CAUSE2 }{ second likely cause} \item{LIK2 }{
likelihood of 2nd cause} \item{CAUSE3 }{ third likely cause} \item{LIK3 }{
likelihood of 3rd cause}
\item{wholeprob }{ full distribution of causes of death}
}
\description{
This function implements the InterVA algorithm specifically for data
collected using the 2022 WHO verbal autopsy instrument.  It
produces individual cause of death (COD) and population cause-specific mortality
fractions.  The output is saved in a .csv file specified by user.
The calculation is based on the conditional and prior distribution
of 64 CODs. The function can also  save the full probability distribution
of each individual to file. All information about each individual is
saved to a va class object.
}
\details{
Be careful if the input file does not match InterVA input format strictly.
The function will run normally as long as the number of symptoms are
correct. Any inconsistent symptom names will be printed in console as
warning. If there is a wrong match of symptom from warning, please change 
the input to the correct order.
}
\examples{

data(va2022)

## to get easy-to-read version of causes of death make sure the column
## orders match InterVA2022 standard input this can be monitored by checking
## the warnings of column names

sample.output1 <- InterVA2022(va2022, HIV = "h", Malaria = "l", 
    Covid="v", write = FALSE, directory = tempdir(),
    filename = "VA2022_result", output = "extended", append = FALSE)

\dontrun{
## to get causes of death with group code for further usage
sample.output2 <- InterVA2022(va2022, HIV = "h", Malaria = "l", Covid="v",
    write = FALSE, directory = "VA test", filename = "VA2022_result_wt_code",
    output = "classic", append = FALSE, groupcode = TRUE)
}


}
\seealso{
\code{\link{InterVA2022.plot}}
}
\author{
CDC Foundation
}
\keyword{InterVA}
