% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA2022.summary.R
\name{print.interVA2022_summary}
\alias{print.interVA2022_summary}
\title{Print method for summary of the results obtained from InterVA2022 algorithm}
\usage{
\method{print}{interVA2022_summary}(x, ...)
}
\arguments{
\item{x}{summary of InterVA2022 results}

\item{...}{not used}
}
\description{
This function prints the summary message of the fitted results.
}
\keyword{InterVA}
