% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA2022-package.R
\docType{data}
\name{probbase2022}
\alias{probbase2022}
\title{Conditional probability of the 2022 Probbase}
\format{
A data frame with 343 observations on 71 variables.
}
\description{
This is the table of conditional probabilities of symptoms given CODs, along with
prior probabilities in the first row. The values correspond
to the WHO VA 2022 instrument.
}
\examples{

data(probbase2020)

}
\keyword{datasets}
