% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA2022-package.R
\docType{data}
\name{RandomVA2022}
\alias{RandomVA2022}
\title{23 records of Sample Input}
\format{
23 arbitrary input records.
}
\description{
This is a dataset consisting of 23 arbitrary sample input deaths in the
acceptable format of the WHO2022 data format. Any dataset that needs to be analyzed by this
package should be in the same format. The order of the input fields must
not be changed.
}
\examples{

data(RandomVA2022)

}
\keyword{datasets}
