% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA2022.summary.R
\name{summary.interVA2022}
\alias{summary.interVA2022}
\title{Summary of the results obtained from InterVA2022 algorithm}
\usage{
\method{summary}{interVA2022}(object, top = 5, id = NULL, InterVA.rule = TRUE, ...)
}
\arguments{
\item{object}{fitted object from \code{InterVA2022()}}

\item{top}{number of top CSMF to show}

\item{id}{the ID of a specific death to show}

\item{InterVA.rule}{If it is set to "TRUE", only the top 3 causes reported by
InterVA2022 is calculated into CSMF as in InterVA2022. The rest of probabilities
goes into an extra category "Undetermined". Default set to "TRUE".}

\item{...}{not used}
}
\description{
This function prints the summary message of the fitted results.
}
\examples{
\dontrun{
data(va2022)

## to get easy-to-read version of causes of death make sure the column
## orders match InterVA2022 standard input this can be monitored by checking
## the warnings of column names

sample.output <- InterVA2022(va2022, HIV = "h", Malaria = "l", Covid = "v",
    write = FALSE, directory = tempdir(), filename = "VA2022_result",
    output = "extended", append = FALSE)

summary(sample.output)
summary(sample.output, top = 10)
summary(sample.output, id = " 1")
}
}
\keyword{InterVA}
